Remove the type parameter from the fixture macros
This effectively means that users are required to manually cast
the data argument within the fixture functions, since these functions
are now exposing a `void *` parameter instead of the actual data type
as before, which is a shame.

The main advantage of this change is that it's leading the way to
a more consistent interface regarding how fixtures are defined
and attached to test cases.

More to come in the following commits.