Allow to set a fixture as any other option
This means that there is no need to call a specialized macro like
`RX_TEST_CASE_FIXTURE` anymore and that a test suite configured with
a fixture will have all of its test cases inheriting that same fixture.

On the downside, this further cement the fact that users are now
required to manually cast `void *` args to the fixture's data type.